fix(box): rename file
